revert code cleanup changes
